feat: tratamento de erros e refatoracao de arquivos
