+++ miaudote/src/main/kotlin/com/projeto2/miaudote/application/handler/ProcessorHandler.kt
@@ -1,0 +1,8 @@
+package com.projeto2.miaudote.application.handler
+
+import org.springframework.stereotype.Service
+
+@Service
+abstract class ProcessorHandler<T: RequestHandler> {
+    abstract fun process(handler: T): Result<Any>
+}
