+++ miaudote/src/main/kotlin/com/projeto2/miaudote/application/handler/RequestHandler.kt
@@ -1,0 +1,3 @@
+package com.projeto2.miaudote.application.handler
+
+interface RequestHandler
