--- miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/Adapters/Request/Login.kt
@@ -1,7 +1,0 @@
-package com.projeto2.miaudote.apresentation.Adapters.Request
-
-class LoginRequest(
-    val username: String,
-    val senha: String
-) {
-}
