--- miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/Adapters/Request/UsuarioRequest.kt
@@ -1,12 +1,0 @@
-package com.projeto2.miaudote.apresentation.Adapters.Request
-
-class UsuarioCreate(
-    val nome: String?,
-    val sobrenome: String?,
-    val username: String?,
-    val senha: String?,
-    val email: String?,
-    val cpf: String?,
-    val descricao: String?,
-    val contato: String?,
-)
