--- miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/Adapters/Response/Login.kt
@@ -1,7 +1,0 @@
-package com.projeto2.miaudote.apresentation.Adapters.Response
-
-class LoginResponse(
-    val username: String,
-    val accessToken: String,
-    val expiresIn: Long,
-)
