+++ miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/Request/Login.kt
@@ -1,0 +1,6 @@
+package com.projeto2.miaudote.apresentation.Request
+
+class LoginRequest(
+    val username: String?,
+    val senha: String?
+)
