+++ miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/Request/PetRequest.kt
@@ -1,0 +1,11 @@
+package com.projeto2.miaudote.apresentation.Request
+
+data class PetCreate(
+    val nome: String?,
+    val sexo: String?,
+    val porte: String?,
+    val idade: String?,
+    val tipo: String?,
+    val castrado: String?,
+    val descricao: String?,
+)
