+++ miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/Request/UsuarioRequest.kt
@@ -1,0 +1,12 @@
+package com.projeto2.miaudote.apresentation.Request
+
+class UsuarioCreate(
+    val nome: String?,
+    val sobrenome: String?,
+    val username: String?,
+    val senha: String?,
+    val email: String?,
+    val cpf: String?,
+    val descricao: String?,
+    val contato: String?,
+)
