+++ miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/Response/Login.kt
@@ -1,0 +1,7 @@
+package com.projeto2.miaudote.apresentation.Response
+
+class LoginResponse(
+    val username: String,
+    val accessToken: String,
+    val expiresIn: Long,
+)
