--- miaudote/src/main/kotlin/com/projeto2/miaudote/apresentation/handler/RequestHandler.kt
@@ -1,3 +1,0 @@
-package com.projeto2.miaudote.apresentation.handler
-
-interface RequestHandler
