--- miaudote/src/main/kotlin/com/projeto2/miaudote/enums/Castrado.kt
@@ -1,16 +1,0 @@
-package com.projeto2.miaudote.enums
-
-import com.fasterxml.jackson.annotation.JsonFormat
-
-@JsonFormat(shape = JsonFormat.Shape.OBJECT)
-enum class Castrado(val id: String, val nome: String) {
-    C("C", "Castrado"),
-    N("N", "Não castrado"),
-    ;
-}
-
-fun String.toCastrado(): Castrado? {
-    val castrado = Castrado.entries.find { it.id == this }
-    // castrado ?: error("Informação de castrado não encontrado.")
-    return castrado
-}
