--- miaudote/src/main/kotlin/com/projeto2/miaudote/enums/converters/TipoConverter.kt
@@ -1,17 +1,0 @@
-package com.projeto2.miaudote.enums.converters
-
-import com.projeto2.miaudote.enums.Tipo
-import com.projeto2.miaudote.enums.toTipo
-import jakarta.persistence.AttributeConverter
-import jakarta.persistence.Converter
-
-@Converter(autoApply = true)
-class TipoConverter : AttributeConverter<Tipo?, String?> {
-    override fun convertToDatabaseColumn(tipo: Tipo?): String? {
-        return tipo?.id
-    }
-
-    override fun convertToEntityAttribute(id: String?): Tipo? {
-        return id?.toTipo()
-    }
-}
