--- miaudote/src/main/kotlin/com/projeto2/miaudote/security/CorsConfiguration.kt
@@ -1,15 +1,0 @@
-package com.projeto2.miaudote.security
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class CorsConfiguration : WebMvcConfigurer {
-    @Override
-    override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-            .allowedOrigins("*")
-            .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS", "HEAD", "TRACE", "CONNECT")
-    }
-}
