--- miaudote/src/main/kotlin/com/projeto2/miaudote/security/SecurityConfig.kt
@@ -1,71 +1,0 @@
-package com.projeto2.miaudote.security
-
-import com.nimbusds.jose.jwk.JWK
-import com.nimbusds.jose.jwk.JWKSet
-import com.nimbusds.jose.jwk.RSAKey
-import com.nimbusds.jose.jwk.source.ImmutableJWKSet
-import com.nimbusds.jose.jwk.source.JWKSource
-import com.nimbusds.jose.proc.SecurityContext
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.http.HttpMethod
-import org.springframework.security.config.Customizer
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.http.SessionCreationPolicy
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
-import org.springframework.security.crypto.password.PasswordEncoder
-import org.springframework.security.oauth2.jwt.JwtDecoder
-import org.springframework.security.oauth2.jwt.JwtEncoder
-import org.springframework.security.oauth2.jwt.NimbusJwtDecoder
-import org.springframework.security.oauth2.jwt.NimbusJwtEncoder
-import org.springframework.security.web.SecurityFilterChain
-import java.security.interfaces.RSAPrivateKey
-import java.security.interfaces.RSAPublicKey
-
-
-@Configuration
-@EnableWebSecurity
-class SecurityConfig(
-    @Value("\${jwt.public.key}")
-    private val key: RSAPublicKey,
-
-    @Value("\${jwt.private.key}")
-    private val private: RSAPrivateKey,
-) {
-
-    @Bean
-    fun securityfilterChain(http: HttpSecurity): SecurityFilterChain {
-        http.csrf { it.disable() }.authorizeHttpRequests { auth ->
-            auth.requestMatchers("api/auth/login", "api/usuario/cadastrar", "api/pet/obter-pets").permitAll()
-                .requestMatchers(HttpMethod.OPTIONS, "/**" ).permitAll()
-                .anyRequest().authenticated()
-        }
-            .httpBasic(Customizer.withDefaults())
-            .oauth2ResourceServer { conf ->
-                conf.jwt(Customizer.withDefaults())
-            }
-            .sessionManagement { session ->
-                session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-            }
-        return http.build()
-    }
-
-    @Bean
-    fun jwtDecoder(): JwtDecoder {
-        return NimbusJwtDecoder.withPublicKey(key).build()
-    }
-
-    @Bean
-    fun jwtEncoder(): JwtEncoder {
-        val jwk: JWK = RSAKey.Builder(key).privateKey(private).build()
-        val jwks: JWKSource<SecurityContext> = ImmutableJWKSet(JWKSet(jwk))
-        return NimbusJwtEncoder(jwks)
-    }
-
-    @Bean
-    fun passwordEncoder(): BCryptPasswordEncoder {
-        return BCryptPasswordEncoder()
-    }
-}
